Add support for setting the presence of an account through the plasma services mechanism.
Everything seems to work OK (through plasmaengineexplorer), but its not stress tested, so don't be surprised if you break it.
There are still some smallish issues that need sorting out, but that will require a bit of talk with the telepathy mailing list to clarify some points.
Code cleanup and consistency checking of parameter naming will also come soon.

svn path=/trunk/playground/base/plasma/dataengines/presence/; revision=949339