Part 1 of 2 for integrating the global presence into the presence dataengine
This is the first patch for integrating the global presence into the presence dataengine.
For now, a "GlobalPresence" source is created. This source will keep track of the global presence. For now
there's only the source that reads the presence.
Next step will be creating the jobs to set a desired presence

REVIEW: 102813
Reviewed By: David Edmundson